AN-178 WIP moving to generic embed handler
