+++ includes/apple-exporter/class-component-factory.php
@@ -93,10 +93,6 @@ class Component_Factory {
 		self::register_component( 'facebook', '\\Apple_Exporter\\Components\\Facebook' );
 		self::register_component( 'instagram', '\\Apple_Exporter\\Components\\Instagram' );
 		self::register_component( 'table', '\\Apple_Exporter\\Components\\Table' );
-		self::register_component( 'flickr', '\\Apple_Exporter\\Components\\Flickr' );
-		self::register_component( 'wp-embed', '\\Apple_Exporter\\Components\\WP_Embed' );
-		self::register_component( 'spotify', '\\Apple_Exporter\\Components\\Spotify' );
-		self::register_component( 'soundcloud', '\\Apple_Exporter\\Components\\SoundCloud' );
 		self::register_component( 'iframe', '\\Apple_Exporter\\Components\\Embed_Web_Video' );
 		self::register_component( 'img', '\\Apple_Exporter\\Components\\Image' );
 		self::register_component( 'video', '\\Apple_Exporter\\Components\\Video' );
